% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkmat}
\alias{checkmat}
\title{Plot Adjacency Matrix}
\usage{
checkmat(mat)
}
\arguments{
\item{mat}{matrix to be plotted.}
}
\value{
None.
}
\description{
Produces a simple plot of the supplied adjacency matrix.
}
\examples{
data(mid_mat)
checkmat(mid_mat)

}
