% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_mid-data.R
\docType{data}
\name{mid_edgelist}
\alias{mid_edgelist}
\title{Dyadic Militarized Interstate Dispute (MID) Network (edgelist)}
\format{
An edgelist with 134 entries (i.e. matrix with 2 columns and 134 rows).
}
\source{
Zeev Maoz's website with the original data (no longer working): \url{http://vanity.dss.ucdavis.edu/~maoz/dyadmid.html}
}
\usage{
data(mid_edgelist)
}
\description{
An undirected network of militarized interstate disputes between two countries. This network data is adapted from Zeev Maoz's (2005) dyadic MID data set. The network includes 189 states in the international system during the post-Cold War period (1989-1995) except Samoa. The row and column names are country codes from the Correlates of War project.
}
\references{
Z Maoz, 2005, Dyadic MID Dataset (version 2.0).
}
\keyword{datasets}
