% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_mid-data.R
\docType{data}
\name{mid_node_attr}
\alias{mid_node_attr}
\title{Global Conflict Actor Attributes}
\format{
A data frame with 189 rows and 5 columns:
\describe{
\item{name}{The state's name.}
\item{intrastate}{An indicator for whether the actor is engaged in intrastate conflict.}
\item{dem}{An indicator for whether the actor is a democracy.}
\item{cinc}{The actor's Composite Index of National Capability (CINC) score. All nonstate actors are coded to have a CINC score of zero.}
\item{trans}{An indicator for whether the actor is a transitional regime.}
}
}
\source{
Gleditsch's website: \url{http://ksgleditsch.com/publ.html}

UCDP website: \url{https://ucdp.uu.se/downloads/}
}
\usage{
data(mid_node_attr)
}
\description{
A data frame of six attributes associated with the 189 states in the interstate conflict network for the period of 1989-1995. For details on how temporally dynamic variables are aggregated, see Chen (2021). The row ids are Correlates of War country codes.
}
\references{
Allansson, M, E Melander, and L Themner, 2017, "Organized violence, 1989-2016." \emph{Journal of Peace Research} 54(4): 727-742. \url{https://doi.org/10.1177/0022343317718773}

Chen, THY, 2021, "Statistical Inference for Multilayer Networks in Political Science." \emph{Political Science Research and Methods} 9(2): 380-397. \url{https://doi.org/10.1017/psrm.2019.49}

Gleditsch, KS, I Salehyan, and K Schultz, 2008, "Fighting at home, fighting abroad: How civil wars lead to international disputes." \emph{Journal of Conflict Resolution} 52(4): 479-506. \url{https://doi.org/10.1177/0022002707313305}
}
\keyword{datasets}
